requirements.txt
selenium~=4.1.0
PyYAML~=6.0